Renamed files to use ShowAppMenu instead of Open Menu RPC
