+++ test_scripts/API/ShowAppMenu/002_Show_App_Menu_RPC_without_Sub_MenuID_SUCCESS.lua
@@ -4,16 +4,15 @@
 -- Description:
 -- In case:
 -- 1) Mobile application is set to appropriate HMI level and System Context MENU, MAIN
--- 2) Mobile application is added SubMenu with menuID  = 5
--- 3) Mobile sends ShowAppMenu request with menuID = 5 parameter to SDL
--- 4) SDL sends ShowAppMenu request with menuID parameter to HMI
--- 5) HMI sends ShowAppMenu response with resultCode = SUCCESS to SDL
+-- 2) Mobile sends ShowAppMenu request without menuID parameter to SDL
+-- 3) SDL sends ShowAppMenu request without menuID parameter to HMI
+-- 4) HMI sends ShowAppMenu response with resultCode = SUCCESS to SDL
 -- SDL does:
 -- 1) send ShowAppMenu response with resultCode = SUCCESS to mobile
 ---------------------------------------------------------------------------------------------------
 -- [[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require('test_scripts/API/OpenMenuRPC/commonOpenMenuRPC')
+local common = require('test_scripts/API/ShowAppMenu/commonShowAppMenu')
 
 -- [[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
@@ -27,10 +26,9 @@ runner.Step("App registration", common.registerApp)
 
 runner.Title("Test")
 runner.Step("App activate, HMI SystemContext MAIN", common.activateApp)
-runner.Step("Add menu", common.addSubMenu, { 5 })
-runner.Step("Send show app menu, HMI SystemContext MAIN", common.showAppMenuSuccess, { 5 })
+runner.Step("Send show app menu, HMI SystemContext MAIN", common.showAppMenuSuccess, { nil })
 runner.Step("Set HMI SystemContext to MENU", common.changeHMISystemContext, { "MENU" })
-runner.Step("Show app menu, HMI SystemContext MENU", common.showAppMenuSuccess, { 5 })
+runner.Step("Show app menu, HMI SystemContext MENU", common.showAppMenuSuccess, { nil })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
