+++ test_scripts/API/ShowAppMenu/008_Show_App_Menu_RPC_without_policy_SystemContext_ALERT_DISALLOWED.lua
@@ -12,7 +12,7 @@
 ---------------------------------------------------------------------------------------------------
 -- [[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require('test_scripts/API/OpenMenuRPC/commonOpenMenuRPC')
+local common = require('test_scripts/API/ShowAppMenu/commonShowAppMenu')
 
 -- [[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
