+++ test_scripts/API/ShowAppMenu/010_Show_App_Menu_RPC_different_MenuID_with_Sub_Menu_ID_INVALID_ID.lua
@@ -3,32 +3,34 @@
 -- https://github.com/smartdevicelink/sdl_evolution/blob/master/proposals/0116-open-menu.md
 -- Description:
 -- In case:
--- 1) Mobile application is set to appropriate HMI level and System Context MENU, MAIN
--- 2) Mobile sends ShowAppMenu request without menuID parameter to SDL
--- 3) SDL sends ShowAppMenu request without menuID parameter to HMI
--- 4) HMI sends ShowAppMenu response with resultCode = SUCCESS to SDL
+-- 1) Mobile application is set to FULL HMI level and System Context MAIN
+-- 2) Mobile application is added SubMenu with menuID  = 5
+-- 3) Mobile sends ShowAppMenu request with menuID = 10 parameter to SDL
 -- SDL does:
--- 1) send ShowAppMenu response with resultCode = SUCCESS to mobile
+-- 1) not send ShowAppMenu request to HMI
+-- 2) send ShowAppMenu response to mobile with parameters "success" = false, resultCode = INVALID_ID
 ---------------------------------------------------------------------------------------------------
 -- [[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require('test_scripts/API/OpenMenuRPC/commonOpenMenuRPC')
+local common = require('test_scripts/API/ShowAppMenu/commonShowAppMenu')
 
 -- [[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 config.application1.registerAppInterfaceParams.appHMIType = { "PROJECTION" }
+
+--[[ Local Variables ]]
+local resultCode = "INVALID_ID"
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerApp)
+runner.Step("App activate, HMI SystemContext MAIN", common.activateApp)
 
 runner.Title("Test")
-runner.Step("App activate, HMI SystemContext MAIN", common.activateApp)
-runner.Step("Send show app menu, HMI SystemContext MAIN", common.showAppMenuSuccess, { nil })
-runner.Step("Set HMI SystemContext to MENU", common.changeHMISystemContext, { "MENU" })
-runner.Step("Show app menu, HMI SystemContext MENU", common.showAppMenuSuccess, { nil })
+runner.Step("Add menu", common.addSubMenu, { 5 })
+runner.Step("Send show app menu", common.showAppMenuUnsuccess, { 10, resultCode })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
