+++ test_scripts/API/ShowAppMenu/012_Show_App_Menu_RPC_hmi_Level_FULL_SC_VRSESSION_HMI_OBSCURED_ALERT_REJECTED.lua
@@ -11,7 +11,7 @@
 ---------------------------------------------------------------------------------------------------
 -- [[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require('test_scripts/API/OpenMenuRPC/commonOpenMenuRPC')
+local common = require('test_scripts/API/ShowAppMenu/commonShowAppMenu')
 
 -- [[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
