+++ test_scripts/API/ShowAppMenu/commonShowAppMenu.lua
