--- test_sets/open_menu_rpc.txt
@@ -1,12 +1,0 @@
-./test_scripts/API/OpenMenuRPC/001_Open_Menu_RPC_with_Sub_MenuID_SUCCESS.lua
-./test_scripts/API/OpenMenuRPC/002_Open_Menu_RPC_without_Sub_MenuID_SUCCESS.lua
-./test_scripts/API/OpenMenuRPC/003_Open_Menu_RPC_with_invalid_MenuID_INVALID_ID.lua
-./test_scripts/API/OpenMenuRPC/004_Open_Menu_RPC_without_policy_SystemContext_MAIN_DISALLOWED.lua
-./test_scripts/API/OpenMenuRPC/005_Open_Menu_RPC_without_policy_SystemContext_VRSESSION_DISALLOWED.lua
-./test_scripts/API/OpenMenuRPC/006_Open_Menu_RPC_without_policy_SystemContext_MENU_DISALLOWED.lua
-./test_scripts/API/OpenMenuRPC/007_Open_Menu_RPC_without_policy_SystemContext_HMI_OBSCURED_DISALLOWED.lua
-./test_scripts/API/OpenMenuRPC/008_Open_Menu_RPC_without_policy_SystemContext_ALERT_DISALLOWED.lua
-./test_scripts/API/OpenMenuRPC/009_Open_Menu_RPC_without_HMI_response_GENERIC_ERROR.lua
-./test_scripts/API/OpenMenuRPC/010_Open_Menu_RPC_different_MenuID_with_Sub_Menu_ID_INVALID_ID.lua
-./test_scripts/API/OpenMenuRPC/011_Open_Menu_RPC_with_invalid_data_type_MenuID_INVALID_DATA.lua
-./test_scripts/API/OpenMenuRPC/012_Open_Menu_RPC_hmi_Level_FULL_SC_VRSESSION_HMI_OBSCURED_ALERT_REJECTED.lua
