+++ test_sets/show_app_menu_rpc.txt
@@ -1,0 +1,12 @@
+./test_scripts/API/ShowAppMenu/001_Show_App_Menu_RPC_with_Sub_MenuID_SUCCESS.lua
+./test_scripts/API/ShowAppMenu/002_Show_App_Menu_RPC_without_Sub_MenuID_SUCCESS.lua
+./test_scripts/API/ShowAppMenu/003_Show_App_Menu_RPC_with_invalid_MenuID_INVALID_ID.lua
+./test_scripts/API/ShowAppMenu/004_Show_App_Menu_RPC_without_policy_SystemContext_MAIN_DISALLOWED.lua
+./test_scripts/API/ShowAppMenu/005_Show_App_Menu_RPC_without_policy_SystemContext_VRSESSION_DISALLOWED.lua
+./test_scripts/API/ShowAppMenu/006_Show_App_Menu_RPC_without_policy_SystemContext_MENU_DISALLOWED.lua
+./test_scripts/API/ShowAppMenu/007_Show_App_Menu_RPC_without_policy_SystemContext_HMI_OBSCURED_DISALLOWED.lua
+./test_scripts/API/ShowAppMenu/008_Show_App_Menu_RPC_without_policy_SystemContext_ALERT_DISALLOWED.lua
+./test_scripts/API/ShowAppMenu/009_Show_App_Menu_RPC_without_HMI_response_GENERIC_ERROR.lua
+./test_scripts/API/ShowAppMenu/010_Show_App_Menu_RPC_different_MenuID_with_Sub_Menu_ID_INVALID_ID.lua
+./test_scripts/API/ShowAppMenu/011_Show_App_Menu_RPC_with_invalid_data_type_MenuID_INVALID_DATA.lua
+./test_scripts/API/ShowAppMenu/012_Show_App_Menu_RPC_hmi_Level_FULL_SC_VRSESSION_HMI_OBSCURED_ALERT_REJECTED.lua
